simplify files: Extract cheerioFacade
