+++ src/tableWalker.res.mjs
@@ -1,199 +1,39 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Cheerio from "cheerio";
-import * as Core__Int from "@rescript/core/src/Core__Int.res.mjs";
 import * as Core__Array from "@rescript/core/src/Core__Array.res.mjs";
 import * as Core__Error from "@rescript/core/src/Core__Error.res.mjs";
 import * as Core__Option from "@rescript/core/src/Core__Option.res.mjs";
-import * as Core__Nullable from "@rescript/core/src/Core__Nullable.res.mjs";
+import * as CheerioFacade from "./cheerioFacade.res.mjs";
 import * as Primitive_option from "rescript/lib/es6/Primitive_option.js";
 
-let applyElementToCheerioUnsafe = (function(element, selectorFunction) {
-        return selectorFunction(element)
-      });
-
-function load(html, maybeOptions, isDocument) {
-  let loadedCheerio = Cheerio.load(html, maybeOptions, isDocument);
-  return (unsafeFromJsLand, basicAcceptedElems) => {
-    if (unsafeFromJsLand !== undefined) {
-      return applyElementToCheerioUnsafe(Primitive_option.valFromOption(unsafeFromJsLand), loadedCheerio);
-    } else {
-      return applyElementToCheerioUnsafe(basicAcceptedElems._0, loadedCheerio);
-    }
-  };
-}
-
-function cheerioToElementArray(queriedCheerio) {
-  return queriedCheerio.toArray();
-}
-
-function getParent(element) {
-  return element.parent;
-}
-
-function getParentExn(element) {
-  return Core__Option.getExn(element.parent, undefined);
-}
-
-function getChildren(element) {
-  return element.children;
-}
-
-function getType(element) {
-  return element.type;
-}
-
-function getData(element) {
-  if (element.type !== "text" && element.type !== "comment" && element.type !== "directive") {
-    Core__Error.panic("Trying to get 'data' from non-text element");
-  }
-  return element.data;
-}
-
-function getName(element) {
-  return element.name;
-}
-
-function getColspan(element) {
-  let match = element.type;
-  if (match !== "style" && match !== "tag" && match !== "script") {
-    return Core__Error.panic("Trying to get 'colspan' from non-element node");
-  }
-  let colspan = Core__Nullable.getExn(element.attribs).colspan;
-  if (colspan == null) {
-    return "1";
-  } else {
-    return colspan;
-  }
-}
-
-function getColspanInt(element) {
-  let parsed = Core__Int.fromString(getColspan(element), undefined);
-  if (parsed !== undefined) {
-    return parsed;
-  } else {
-    return 1;
-  }
-}
-
-function getRowspan(element) {
-  let match = element.type;
-  if (match !== "style" && match !== "tag" && match !== "script") {
-    return Core__Error.panic("Trying to get 'rowspan' from non-element node");
-  }
-  let rowspan = Core__Nullable.getExn(element.attribs).rowspan;
-  if (rowspan == null) {
-    return "1";
-  } else {
-    return rowspan;
-  }
-}
-
-function getRowspanInt(element) {
-  let parsed = Core__Int.fromString(getRowspan(element), undefined);
-  if (parsed !== undefined) {
-    return parsed;
-  } else {
-    return 1;
-  }
-}
-
-function getText(element, loadedCheerio) {
-  return loadedCheerio(undefined, {
-    TAG: "AnyNode",
-    _0: element
-  }).text();
-}
-
-function getLengthString(queriedCheerio) {
-  return queriedCheerio.length.toString();
-}
-
-function each(queriedCheerio, callback) {
-  queriedCheerio.each(callback);
-}
-
-function find(queriedCheerio, queryString) {
-  return queriedCheerio.find(queryString);
-}
-
-function remove(queriedCheerio) {
-  return queriedCheerio.remove();
-}
-
-function contents(queriedCheerio) {
-  return queriedCheerio.contents();
-}
-
-function getLast(queriedCheerio) {
-  return queriedCheerio.last();
-}
-
-function getSrc(element) {
-  return Primitive_option.fromNullable(element.attr("src"));
-}
-
-let CheerioFacade = {
-  applyElementToCheerioUnsafe: applyElementToCheerioUnsafe,
-  load: load,
-  cheerioToElementArray: cheerioToElementArray,
-  getParent: getParent,
-  getParentExn: getParentExn,
-  getChildren: getChildren,
-  getType: getType,
-  getData: getData,
-  getName: getName,
-  getColspan: getColspan,
-  getColspanInt: getColspanInt,
-  getRowspan: getRowspan,
-  getRowspanInt: getRowspanInt,
-  getText: getText,
-  getLengthString: getLengthString,
-  each: each,
-  find: find,
-  remove: remove,
-  contents: contents,
-  getLast: getLast,
-  getSrc: getSrc
-};
-
-function _loadCheerio(html) {
+function loadCheerio(html) {
   if (html.indexOf("<html>") !== -1) {
-    return load(html, null, true);
+    return CheerioFacade.load(html, null, true);
   } else {
-    return load(html, null, false);
+    return CheerioFacade.load(html, null, false);
   }
 }
 
 function _sanityCheckHeaders(cheerioWithHeaders) {
-  let headerElements = cheerioWithHeaders.toArray();
+  let headerElements = CheerioFacade.cheerioToElementArray(cheerioWithHeaders);
   let firstHeader = Core__Option.getExn(headerElements[0], undefined);
-  if (firstHeader.name !== "th") {
+  if (CheerioFacade.getName(firstHeader) !== "th") {
     console.log(firstHeader);
     Core__Error.panic("Queried cheerio was expected to contain any headers. But it doesn't");
   }
-  let parent = Core__Option.getExn(firstHeader.parent, undefined);
-  let maybeInvalidHeader = headerElements.find(header => Core__Option.getExn(header.parent, undefined) !== parent);
+  let parent = CheerioFacade.getParentExn(firstHeader);
+  let maybeInvalidHeader = headerElements.find(header => CheerioFacade.getParentExn(header) !== parent);
   if (maybeInvalidHeader === undefined) {
     return;
   }
   let invalidHeader = Primitive_option.valFromOption(maybeInvalidHeader);
-  let thText = invalidHeader.children.map(getData).toString();
-  let expected = parent.name;
-  let actual = Core__Option.getExn(invalidHeader.parent, undefined).name;
+  let thText = CheerioFacade.getChildren(invalidHeader).map(CheerioFacade.getData).toString();
+  let expected = CheerioFacade.getName(parent);
+  let actual = CheerioFacade.getName(Core__Option.getExn(CheerioFacade.getParent(invalidHeader), undefined));
   Core__Error.panic("Multiple headline rows--that's bad. <th>" + thText + "</th>. Expected parent element " + expected + ", but was " + actual);
 }
 
-function parseIntOr(maybeString, fallback) {
-  let n = Core__Int.fromString(maybeString, undefined);
-  if (n !== undefined) {
-    return n;
-  } else {
-    return fallback;
-  }
-}
-
-function removeInnerWhiteSpace(text) {
+function _removeInnerWhiteSpace(text) {
   return text.replace(/\s+/g, " ").trim();
 }
 
@@ -203,14 +43,14 @@ function _getHeaderCells(loadedCheerio) {
     _0: "th"
   });
   _sanityCheckHeaders(ths);
-  let thElements = ths.toArray();
-  let columnCount = Core__Array.reduce(thElements.map(getColspanInt), 0, (prev, curr) => prev + curr | 0);
+  let thElements = CheerioFacade.cheerioToElementArray(ths);
+  let columnCount = Core__Array.reduce(thElements.map(CheerioFacade.getColspanInt), 0, (prev, curr) => prev + curr | 0);
   console.log("Found " + thElements.length.toString() + " table headers (spanning " + columnCount.toString() + " columns).");
   return Core__Array.reduceWithIndex(thElements, [], (headers, header, headerIdx) => {
     let maybePreviousCol = Core__Option.map(headers[headerIdx - 1 | 0], h => h.colEnd);
     let colStart = maybePreviousCol !== undefined ? maybePreviousCol + 1 | 0 : 0;
-    let colEnd = (colStart + getColspanInt(header) | 0) - 1 | 0;
-    let linesOfText = [removeInnerWhiteSpace(getText(header, loadedCheerio))];
+    let colEnd = (colStart + CheerioFacade.getColspanInt(header) | 0) - 1 | 0;
+    let linesOfText = [_removeInnerWhiteSpace(CheerioFacade.getText(header, loadedCheerio))];
     return headers.concat([{
         colStart: colStart,
         colEnd: colEnd,
@@ -219,7 +59,7 @@ function _getHeaderCells(loadedCheerio) {
   });
 }
 
-function getStartIndexForCell(allCells, initialColumnIdx, rowIndex) {
+function _getStartIndexForCell(allCells, initialColumnIdx, rowIndex) {
   let _columnIdx = initialColumnIdx;
   while (true) {
     let columnIdx = _columnIdx;
@@ -244,21 +84,20 @@ function _getDataCells(cheerio) {
     TAG: "StringSelector",
     _0: "tr:has(td)"
   });
-  console.log("Found " + rows.length.toString() + " rows (not including rowspans).");
+  console.log("Found " + CheerioFacade.getLengthString(rows) + " rows (not including rowspans).");
   let allCells = [];
-  rows.each((rowIndex, row) => {
-    let queriedCheerio = cheerio(undefined, {
+  CheerioFacade.each(rows, (rowIndex, row) => {
+    let dataCells = CheerioFacade.find(cheerio(undefined, {
       TAG: "AnyNode",
       _0: row
-    });
-    let dataCells = queriedCheerio.find("td");
+    }), "td");
     let columnIdx = {
       contents: 0
     };
-    dataCells.each((param, cell) => {
-      let colStart = getStartIndexForCell(allCells, columnIdx.contents, rowIndex);
-      let colEnd = (colStart + getColspanInt(cell) | 0) - 1 | 0;
-      let rowEnd = (rowIndex + getRowspanInt(cell) | 0) - 1 | 0;
+    CheerioFacade.each(dataCells, (param, cell) => {
+      let colStart = _getStartIndexForCell(allCells, columnIdx.contents, rowIndex);
+      let colEnd = (colStart + CheerioFacade.getColspanInt(cell) | 0) - 1 | 0;
+      let rowEnd = (rowIndex + CheerioFacade.getRowspanInt(cell) | 0) - 1 | 0;
       allCells.push({
         colStart: colStart,
         colEnd: colEnd,
@@ -272,7 +111,7 @@ function _getDataCells(cheerio) {
   return allCells;
 }
 
-function concatenate(a, b) {
+function _concatenate(a, b) {
   let isPunctuation = c => {
     if (c === "." || c === "," || c === "?" || c === "!" || c === ":") {
       return true;
@@ -292,36 +131,34 @@ function concatenate(a, b) {
 
 function removeInvisibleSourceLineBreaks(cheerio, node) {
   let lines = [];
-  let queriedCheerio = cheerio(undefined, {
+  let nodes = CheerioFacade.cheerioToElementArray(CheerioFacade.contents(cheerio(undefined, {
     TAG: "AnyNode",
     _0: node
-  });
-  let queriedCheerio$1 = queriedCheerio.contents();
-  let nodes = queriedCheerio$1.toArray();
+  })));
   let currentInlineText = {
     contents: ""
   };
   let flushInlineText = () => {
     if (currentInlineText.contents !== "") {
-      lines.push(removeInnerWhiteSpace(currentInlineText.contents));
+      lines.push(_removeInnerWhiteSpace(currentInlineText.contents));
       currentInlineText.contents = "";
       return;
     }
     
   };
   nodes.forEach(node => {
-    let text = getText(node, cheerio).trim();
-    let nodeName = node.name;
+    let text = CheerioFacade.getText(node, cheerio).trim();
+    let nodeName = CheerioFacade.getName(node);
     if (nodeName === "p" || nodeName === "br") {
       flushInlineText();
       if (nodeName === "p" && text !== "") {
-        lines.push(removeInnerWhiteSpace(text));
+        lines.push(_removeInnerWhiteSpace(text));
         return;
       } else {
         return;
       }
     } else if (text !== "") {
-      currentInlineText.contents = concatenate(currentInlineText.contents, text);
+      currentInlineText.contents = _concatenate(currentInlineText.contents, text);
       return;
     } else {
       return;
@@ -332,29 +169,23 @@ function removeInvisibleSourceLineBreaks(cheerio, node) {
 }
 
 function _extractTextFromCell(cheerio, cell) {
-  let queriedCheerio = cheerio(undefined, {
+  CheerioFacade.remove(CheerioFacade.find(cheerio(undefined, {
     TAG: "AnyNode",
     _0: Core__Option.getExn(cell._cheerioEl, undefined)
-  });
-  let queriedCheerio$1 = queriedCheerio.find("small");
-  queriedCheerio$1.remove();
-  let queriedCheerio$2 = cheerio(undefined, {
+  }), "small"));
+  CheerioFacade.remove(CheerioFacade.find(cheerio(undefined, {
     TAG: "AnyNode",
     _0: Core__Option.getExn(cell._cheerioEl, undefined)
-  });
-  let queriedCheerio$3 = queriedCheerio$2.find("sup");
-  queriedCheerio$3.remove();
+  }), "sup"));
   return removeInvisibleSourceLineBreaks(cheerio, Core__Option.getExn(cell._cheerioEl, undefined));
 }
 
 function _extractAndResizeImageUrl(cheerio, cell) {
-  let queriedCheerio = cheerio(undefined, {
+  let imageElement = CheerioFacade.getLast(CheerioFacade.find(cheerio(undefined, {
     TAG: "AnyNode",
     _0: Core__Option.getExn(cell._cheerioEl, undefined)
-  });
-  let queriedCheerio$1 = queriedCheerio.find("img");
-  let imageElement = queriedCheerio$1.last();
-  return Core__Option.map(Primitive_option.fromNullable(imageElement.attr("src")), src => {
+  }), "img"));
+  return Core__Option.map(CheerioFacade.getSrc(imageElement), src => {
     let parts = src.split("/");
     let filtered = parts.filter((param, index) => index !== (parts.length - 1 | 0));
     let lastPart = Core__Option.getExn(filtered[filtered.length - 1 | 0], undefined);
@@ -388,7 +219,7 @@ function _cellHasContent(cell) {
 }
 
 function tableWalker(html) {
-  let cheerio = _loadCheerio(html);
+  let cheerio = loadCheerio(html);
   let headerCells = _getHeaderCells(cheerio);
   let dataCells = _getDataCells(cheerio);
   return dataCells.map(cell => {
@@ -408,15 +239,13 @@ function tableWalker(html) {
 }
 
 export {
-  CheerioFacade,
-  _loadCheerio,
+  loadCheerio,
   _sanityCheckHeaders,
-  parseIntOr,
-  removeInnerWhiteSpace,
+  _removeInnerWhiteSpace,
   _getHeaderCells,
-  getStartIndexForCell,
+  _getStartIndexForCell,
   _getDataCells,
-  concatenate,
+  _concatenate,
   removeInvisibleSourceLineBreaks,
   _extractTextFromCell,
   _extractAndResizeImageUrl,
@@ -424,4 +253,4 @@ export {
   _cellHasContent,
   tableWalker,
 }
-/* cheerio Not a pure module */
+/* CheerioFacade Not a pure module */
